requirements.txt
pycryptodome
requests
bs4
qrcode